requirements.txt
openai
streamlit
pymupdf
faiss-cpu
langchain-huggingface
numpy